+++ src/digital-pass/digital-pass.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { DigitalPassController } from './digital-pass.controller';
+import { DigitalPassService } from './digital-pass.service';
+
+@Module({
+  controllers: [DigitalPassController],
+  providers: [DigitalPassService],
+  exports: [DigitalPassService],
+})
+export class DigitalPassModule {}